+++ flutter-hms-ads/lib/src/adslite/reward/reward_verify_config.dart
@@ -14,10 +14,21 @@
     limitations under the License.
 */
 
-part of huawei_ads;
+part of '../../../huawei_ads.dart';
 
 class RewardVerifyConfig {
+  /// User ID.
+  ///
+  /// `NOTE:`
+  /// This is a user ID of a rewarded ad. This parameter can be combined with a URL and then the URL is encoded.
+  /// In this case, the URL length cannot exceed 1024 bytes. Otherwise, the server-side verification is affected.
   String? userId;
+
+  /// Custom data.
+  ///
+  /// `NOTE:`
+  /// This is a custom parameter of a rewarded ad. This parameter can be combined with a URL and then the URL is encoded.
+  /// In this case, the URL length cannot exceed 1024 bytes. Otherwise, the server-side verification is affected.
   String? data;
 
   RewardVerifyConfig({